+++ package.json
@@ -1,6 +1,6 @@
 {
-  "name": "sublink-plus",
-  "version": "0.0.1",
+  "name": "sublink-worker",
+  "version": "2.0.0",
   "private": true,
   "scripts": {
     "deploy": "npm run setup-kv && wrangler deploy",
