+++ src/SingboxConfigBuilder.js
@@ -1,6 +1,6 @@
-import { SING_BOX_CONFIG, generateRuleSets, generateRules, getOutbounds, PREDEFINED_RULE_SETS } from './config.js';
+import { SING_BOX_CONFIG, generateRuleSets, generateRules, getOutbounds, PREDEFINED_RULE_SETS } from './config/index.js';
 import { BaseConfigBuilder } from './BaseConfigBuilder.js';
-import { DeepCopy, parseCountryFromNodeName } from './utils.js';
+import { DeepCopy, groupProxiesByCountry } from './utils.js';
 
 export class SingboxConfigBuilder extends BaseConfigBuilder {
     constructor(inputString, selectedRules, customRules, baseConfig, lang, userAgent, groupByCountry = false, enableClashUI = false, externalController, externalUiDownloadUrl) {
@@ -145,17 +145,8 @@ export class SingboxConfigBuilder extends BaseConfigBuilder {
 
     addCountryGroups() {
         const proxies = this.getProxies();
-        const countryGroups = {};
-
-        proxies.forEach(proxy => {
-            const countryInfo = parseCountryFromNodeName(proxy?.tag || '');
-            if (countryInfo) {
-                const { name } = countryInfo;
-                if (!countryGroups[name]) {
-                    countryGroups[name] = { ...countryInfo, proxies: [] };
-                }
-                countryGroups[name].proxies.push(proxy.tag);
-            }
+        const countryGroups = groupProxiesByCountry(proxies, {
+            getName: proxy => this.getProxyName(proxy)
         });
 
         const normalize = (s) => typeof s === 'string' ? s.trim() : s;
