+++ test/country-group.test.js
@@ -69,4 +69,18 @@ vmess://ewogICJ2IjogIjIiLAogICJwcyI6ICJ0dzEubm9kZS5jb20iLAogICJhZGQiOiAidHcxLm5v
             expect(missing).toEqual([]);
         }
     });
+
+	it('groupProxiesByCountry helper normalizes names', () => {
+		const sample = [
+			{ name: 'HK-Node-1' },
+			{ tag: '香港节点2' },
+			'US-Node-1 = ss, example.com, 443',
+			'台湾节点 = trojan, example.com, 443'
+		];
+		const grouped = groupProxiesByCountry(sample);
+		expect(Object.keys(grouped)).toEqual(expect.arrayContaining(['Hong Kong', 'United States', 'Taiwan']));
+		expect(grouped['Hong Kong'].proxies).toHaveLength(2);
+		expect(grouped['United States'].proxies).toHaveLength(1);
+		expect(grouped['Taiwan'].proxies).toHaveLength(1);
+	});
 });
